chore: move update code to the externally available path
To use it in Theila update checks.
